Fix lifecycle bug when stopping NeoDataSource
If the checkpointing step fails in NeoDataSource.stop(), we never call
life.shutdown(). Hence we'll never unmap the store files from the page
cache. So later when we try to close the page cache it will fail due
to still mapped files.  This change we'll make sure that we always
shutdown all the lifecycled instances in the internal life in
NeoDataSource.