Add config option for default `allowed` value
When procedures and UDFs come from a third party, where modifying source code
is not possible, it was previously not possible to assign an `allowed` role to them.
This commit introduces a new config setting that may be used to set a specific role
to any procedures and UDFs that do not have their `allowed` annotation set.