Limit how much of the store file that we stream.
At the moment when a follower/read replica gets sent
store files we tell it how many bytes it should expect
to receive.

If the file size changes during the copy process we can
end up sending more bytes than we promised and the
client doesn't handle that very well and story copy fails.

With this commit we introduce our own implementation of
ReadableByteChannel which will only send the portion of the
file that we promised when we first determined its size.