Fix eagerness bug on property reads in the horizon
Previous to this commit, the eagerness analysis system has been working with
the assumption that the `QueryGraph` encodes all reads being done in a
`PlannerQuery`. This is not correct, the `QueryHorizon` may also encode reads.
This commit fixes eagerness issues related to property reads in the horizon,
when those conflict with writes being done in the `QueryGraph`.

Still left to do are reads made by functions in the horizon (`labels()`, `properties()`).
Additionally, potential conflicts between reads in horizon and subsequent
writes in the tail(s) of the PQs.

Also fixed for the rule planner.