Add `version` command to `neo4j-admin`
This command can be pointed against a store and reports information
about its version. Two example outputs are

- If store matches latest known version to the tool:

```
Store version:      v0.A.7
Introduced in:      3.0.0
```

- If store is older than latest known version to the tool:

```
Store version:      v0.A.5
Introduced in:      2.2.0
Superceded in:      2.3.0
```

Help text is as follows:

```
usage: neo4j-admin version --store=<path-to-dir>

Checks the version of a Neo4j database store. Note that this command expects a
path to a store directory, for example --store=data/databases/graph.db.

options:
  --store=<path-to-dir>   Path to database store to check version of.
```