Fail loudly when doing backup against an old version backup
Incremental backup has the following scenarios, when run against an
existing backup B:

1. B is "up to date" (incremental backup possible)
2. B is "too far out of date" (full backup required)
3. B is a backup of an older database store (because you just updated
   to 3.1 and your backup is of 3.0)
4. B is a backup of a different store but the same version

Previously, 3 and 4 were intended to behave differently.

4 would correctly fail with an error to avoid deleting someone's
precious backup files.

3 failed with an error due a bug in the error handling logic, the
intention was to move the old backup and then do a full
backup. However, another bug (masked by the error handling bug) would
have deleted that directory.

Regardless, the correct thing is to fail for 3. The primary reason
being that it is impossible (or unreasonably hard) to determine if
both 3 and 4 apply (e.g., it's a backup of a different store which
also has a different version). This because we start an embedded
database server as part of the backup, and that will always fail if
the store is of a different version.

Conflicts:
	enterprise/backup/src/main/java/org/neo4j/backup/BackupTool.java