Fix shutdown race issue with store copy
When recovering a temporary database copied from another machine, we
borrow the pagecache from the current database.  If the database is
shutdown before the recovery completes the page cache shutdown fails
due to file still mapped from the temporary database. This change
fixes it by abortint recovery if the database is already shutdown, or
the shutdown waits for the recovery to complete in order to cleanly
shutdown the page cache.