Making Bolt advertised address affined to HTTP host
This is a belt-n-braces style commit that tries to address some of the
inconveniences of Bolt server being a kernel extension.
If no sane advertised address is specified for Bolt, then the discovery
service will now try to figure one out using the HTTP Host header
as the basis for computing a Bolt connection URI.

This embeds some assumptions around advertised addresses that may
not be true in all possible cases but should be true for the vast majority.
In particular if the server is bound to a different NIC for HTTP and Bolt,
it's likely this won't work as expected.