Fix a shutdown issue in causal clustering
Sometimes when you shutdown the cluster and the leader is trying to
replicate a transaction, it might happen that the replication effort
is failing since the followers are already gone and the transaction
therad is stuck in the replication loop forever.

This commit will fix that by making the replication code aware of when
the database is shutdown.