DumpLogicalLog can dump transactions relevant to cc report
after a consisistency checker has run a report is made, potentially including
incosistencies. DumpLogicalLog can be given a --ccfilter which should
point to such a report file. DumpLogicalLog will then only include transactions
which touches any of the records in the cc report and mark the records it
matched with a `<----` arrow to easier navigate.