Fix bug on close in NeoStores
If any store does not close cleanly but throws an exception, the
subsequents stores in store array are not gonna be closed.  This will
make the page cache to fail when shutdown due to still store files
mapped in it.  This change will make sure we attempt to close all the
stores even if exceptions are thrown.