Record MapProjection scopes to avoid rewrite conflicts
MapProjection needs the current scope, so it's been recorded in the semantic state.
But since MapProjections can contain other AST elements that get rewritten, it's important to
put this information in a rewrite-safe format and not keep it in the SemanticState, where any
rewrite will make the information hard to get at.

This commit puts the scope information into the MapProjection AST-object.