Copy stable node on write
When writing key content to a node, a new version of that node is created with generation equal to unstableGeneration, if node does not already have generation = unstableGeneration. If that is the case, changes are done directly in that node.
When only updating sibling, newGen or child pointers, a new version is not created as generation of those pointers are already handled by GSPP.

SplitResult -> StructurePropagation is now passed in from SingleIndexWriter instead of returned in method calls.
It is updated with information about new generation of nodes being created.

Various javadoc cleanup.