GSPP - Read generation together with pointer
When reading a pointer, sometimes we need to be able to retrieve the generation
of that pointer. We let GSPP store information about how that information can be retrieved
in the most significant 2 bytes of the pointer, as only 6 bytes are used for the
actual page id.

Since GSPP#read now gives more information in the returned long than just
the pointer, callers who receive read results can no longer treat them as
purely pointers.

This is mostly a problem in tests which asserts pointers or otherwise
deals with read results directly.
For the main code this change is pretty small since the code that hands
out these read results with gen information added will also be aware
of read results as arguments, knowing that it's only the low 6B that
represent pointer.

Encoding generation and pointer in a single long avoids garbage.

Use case example:
TreeNode treeNode = ...
long readResult = GenSafePointerPair.read( cursor, stableGen, unstableGen, logicalPos );
if ( GenSafePointerPair.isSuccess( readResult ) )
{
    long pointer = GenSafePointerPair.pointer( readResult );
    long gen = treeNode.pointerGen( cursor, readResult );
}
else
{
    // Failure encoded as before
}