Flushes label updates before schema changes in batch applier
when making a schema change, e.g. creating an index it may be that the
label scan store is selected to drive the index population. If so then
all label updates needs to be in the label scan store, otherwise they
will be missed.

Previously only index updates were flushed before schema changes,
this commit also flushes label changes when this happens so that index
population will see all updates.