Extra checks reading nodeType and keyCount from tree nodes
Becuse it doesn't make sense to read anything else after reading nodeType
and seing that it isn't even a tree node.

Also keyCount is affecting other reads, which if read from a free-list node
(which is possible due to id-reuse), may be larger than the size of a node and
then cause that next other read to go out-of-bounds in a PageCursor,
which in the end would cause the out-of-bounds check after do-shouldRetry loop
to throw exception.

Now the read key counts are checked for sanity and treated just like reading
from node of a different type.