InternalTreeLogic uses manual stack instead of recursion
this reduces number of page pins since the whole depth won't need to
be traversed up and down for every insert/remove, e.g:

````
                    [A]
             ------/ | \------
            /        |        \
          [B]       [C]       [D]
         / | \     / | \     / | \
       [E][F][G] [H][I][J] [K][L][M]
````

where previously an insert into F, then G, then L would result in page pins
(every pin also including a binary key search):

- for F: A, B, F, B, A
- for G: A, B, G, B, A
- for L: A, D, L, D, A
= totaling 15 pins

but now instead this would look like:

- for F: A, B, F
- for G: B, G
- for L: A, D, L
= totaling 8 pins

the more co-located updates are within the same session (i.e. for the same writer)
the fewer pins and binary searches are required. This works by keeping a stack of where
the writer is in the tree and the key ranges for every level it has gone down.
For every insert/remove the current location is asked whether or not it covers the key
to insert/remove, if not then stack is popped until finding a level which covers the key.

This stack consists of an array of objects initialized once when the GBPTree
instance is constructed and the stack is just an integer knowing which array item
is the current one, i.e. it's garbage-free. Popping the stack also doesn't involve
page pins since the stack knows the key ranges.

tl;dr
The net result of this change is fewer page pins and binary searches to find
correct leaf to perform inserts/removals at. Worst-case is slightly better than before
and best-case is very much better - sequential batch insert being a couple of times
faster due to magnities fewer pins and binary searches.