fix seeding new member from idle cluster backup
Problem is that using a backup of an idle cluster will not require
pulling the last transaction it contains, so it used to try to
pull the next transaction which does not exist yet since there is
no writing going on. We now instead pull starting from one
additional transaction back.

A store bootstrapped in a causal cluster always has a null
transaction added and only in a really misconfigured system
with pruning so aggressive it prunes everything could a
situation exist where there is no transaction available to
pull. Even if this happened it would not be harmful, the
member would just fail to join. In such a case the system
should fix its configuration.

So in conclusion this strategy should always work.