SeekCursor now traverse to correct leaf when seeking backward in search for prim key
Previously when seeking for a range starting on a key equal to a key in an internal node,
a prim key, SeekCursor would always land on a leaf to the left of the correct one when
seeking backwards.
This would be made up for by the SeekCursor by moving one node to the right to find the actual
start of the range before starting returning results to user. Thus this bug did not cause any
wrong results to occur.