Added new config option: dbms.config.strict_validation
By default this setting is false, for now. It causes config parsing to
throw errors on unknown settings (in our namespace).

Settings outside our namespaces are still retained and not warned
about.

Also changed so that warning is a flag, and it is only true when
reading the config file (so just once).