Fix bug with projection in compiled runtime
Since the removal of the distinction between variables and projections in
the CodeGenContext, adding new variables for existing query variable names
to the context in a projection could cause code generation of earlier
clauses to pick the wrong variable name.

This fix assumes that for a projection we do not need to add new variables
or extra projection code for query variables that already exists in the context.