Trace causes of page cursor errors from file closed
Closing a paged file will now record its stack trace in an exception, and if
a page cursor runs into trouble because it is operating on a file that has been
unmapped, then the IllegalStateException will now also contain the file closed
trace exception as a suppressed exception. This allows us to see where and why
a file was closed, and give us clues as to why we still have a cursor on such
a file.