Adds test for GBPTree which auto-detects format changes
this test works by having a committed current-format.zip in src/test/resources
and on failure reading and consistency checking that tree notifies developer that
a FORMAT_VERSION needs to be incremented as well. When developer has incremented
this version a new current-format.zip is generated and must be committed together
with the other changes which changed the format.