Use TestDirectory for the StoreMigrationIT
In order to make this work, the TestDirectory was opened up a bit, so it can
be used to prepare a test directory, without being properly initialised as a
TestRule. The reason for this is that we needed to use it from the parameter
generating method of a @Parameterised test, and that is called before even
the initialisation of @ClassRules.