Added native systemd service
and moved the sysvinit service so it is installed by dh_installinit.

dh_installinit will do the right thing depending on the system you are
installing on. If the system has systemd, it will enable the native
systemd unit. If it has sysvinit, it will enable the old sysvinit
service.

I have also made sure that the service will be called `neo4j` for both
the neo4j and neo4j-enterprise packages.