Added a way to override default values in config for docs
Some settings are calculated at runtime and depend on the current
machine. This is problematic for documentation which uses the default
values and prints them in the manual. This change adds an Annotation
which can specify what to  put in the manual if the actual default
value is not suitable.

Also added the necessary plumbing to access the information in the
annotations via the Config class.