Fix PageCursor being closed twice causing shared reference between PageCursors
The PageCursors are pooled in a linked list (in CursorPool). When a PageCursor is
closed they are added to this list. Closing a PageCursor twice would add the same
object to this list twice. Resulting in subsequent calls to io to return the same object.

This was also visible when opening a new linked cursor after already closing the previous
linked cursor.