Better and clearer lifecycle parallelization in ClusterManager
Introducing ParallelLifecycle which performs lifecycle status changes
on its members in parallel.

ParallelLifecycle is used for shutting down databases in parallel at
end of test, this to reduce time spent there.

A bunch of ad-hoc code related to parallelization of lifecycle of dbs
is removed in favor of ParallelLifecycle.

This also required a change in how free ports are discovered.
Port allocation is now Faster by keeping a memory of discovered free ports
and allocating and freeing those between tests. This takes time off of every
HA test since it takes roughly 1 sec to discover a free port.