LuceneLabelScanStore gets told if it's read-only
previously it was always writable, no matter what the neo4j config said.
Integerestingly compliance testing tested the LuceneLabelScanStore directly
and it asserted that a read-only label scan store shouldn't be able to
create an initial store.

When looking at the rest of the stores, they are actually created, even when
read-only mode is set. Label scan store just happened to not fail when it was
in read-only mode. So now lss aligns with stores in general in that they can
be created even in read-only mode.