Attempt to stabilize TransactionConstraintsIT
It sets up a cluster of 3 where one of the members is a slave-only.
Inconveniently the slave-only member was the first one, which usually creates
the cluster and becomes the master. Because of this there may have been a couple
of more round-trips and unnecessary re-elections causing join timeout.

Changed the slave-only instance to be the second instance in the cluster.