Sets native label scan store as default
At the same time changes the label scan store setting

- Renames it to dbms.label_index Because to the user it's an index,
  an internal setting though since normally there's no need to change it.
- Sets a default value "native" for that setting. This is because we want
  to pin down which implementation is used instead of selecting highest prioritized
  every time, which may result in different implementations - depending on environement.

Also removed InMemoryLabelScanStore. This was mostly there because the main implementation
of the label scan store lived in a module "above" kernel and so kernel tests needed its
own for testing a full database. Now the native label scan store lives inside kernel and so
the in-memory test version is redundant.