filter out data directory
We support deployments, for example in test or embedded, where
instead of the store directory being a subdirectory of the
data directory, in fact the inverse is true with data and
hence data/cluster-state being a subdirectory of the store
directory. This is a legacy thing and not optimal. This commit is
just a quick-fix for the recent breakage due to the move of
cluster-state into the data directory.