Attempt fix for ClusterManager#startCluster() race
A race where the normal methods like allSeesAllAsAvailable()
would sometimes return true if there were no db started yet and therefore
return prematurely. The test would then continue and request a db
with a certain role, wouldn't find it and would fail because of it.

This issue wasn't observed locally when coming up with the fix, but based
on the logs on existing failing builds it looked like this could be the cause.