Unflake TestBranchedData
Problem being that after a role switch there were 3 big transactions w/ indexing and everything
taking place which, on some JVMs, caused full GC pauses on up to 2s. This was a problem because
the heartbeat timeout in ClusterManager is 2s and so the cluster would go into an unstable state
and try to switch role right when the transaction would commit, causing the test to fail.

This commit changes this test so that it no longer performs these big transactions.
There were no apparent reason for those transactions being so big, potentially it could be
(as the comment in the test suggests) that it wanted the lucene index files to change underneath.
Although there were no checking of those files, rather it would most likely use the suggested
file changes to verify that a branch + store-copy was actually performed.

Now instead the test installs a monitor to assert that a store-copy is performed on the
correct instances after the repair has been done.