More consistent DynamicIndexStoreViewIT test time
This test takes a couple of seconds to run normally, but now and then takes
up to minutes or hours!

Problem was that there was a race where five threads were created, threads which
busy-looped on making transactions revolving nodes and properties.
Such transactions acquire the schema read lock. The main test thread would,
after starting these other threads, go on to creating an index, a transaction
which needs to acquire the schema write lock. If those other threads had started
before the main thread would have came as far as acquiring the schema write lock
the test would get itself into a situation where the main thread could face
starvation and be waiting for an unknown amount of time since the other threads
would constantly acquire the schema read lock, overlapping with each other
and basically never let the main thread create the index.

This commit starts the other threads, but let them wait for a count down latch
and let them loose as soon as the index has been created, letting them race
with the population of that index. Number of concurrent updates fed into the
index populator has been verified to be the same as before this change.