Cleanup error logging in Bolt
This commit removes publishing policy from error classification. It was useful
when we tried to send some errors to the user and only log other errors.
Currently we try to send all errors to the user and log only database errors.
So there is no need to publishing policy because we just need to know if error
is a database error to log it.

Also removed self reference to `debug.log` from error messages. It was
previously possible for messages like "[o.n.b.v.r.ErrorReporter] Client
triggered an unexpected error [UnknownError]: null. See debug.log for more
details..." to appear in `debug.log`. This was caused by the fact that we
duplicate all user log messaged to internal log. See `StoreLogService` and `DuplicatingLogProvider` for more details.