Compute sleep time for dbms.listQueries
This restructures the code to capture a snapshot state of
ExecutingQuery. The purpose of this is to ensure that all different time
fields for a single query are computed based on the same point in time.
This should avoid cases where the sum of cpuTime and waitTime could be
greater than elapsedTime due to the fact that they are captured at
different points in time.