cleanup and clarify around topology/discovery
The topology service is an internal service provided by the discovery subsystem,
also referred to as the discovery service.

The TopologyService provides a read-only interface for reading the eventually
consistent topology-related information, while the CoreTopologyService extends
upon that with additional methods required in special circumstances which
require support from the underlying discovery service.

The distinction was not clear from the naming nor documented. This commit
is a first step towards better clarity.

ReadReplicaTopologyService was renamed simply to TopologyService since the
distinction is not on who the client of the service is, but rather on
the width of the interface required. There are clients of TopologyService
both in cores and read replicas, and it can give the complete topology, thus
naming it with ReadReplica makes little sense.