Creating uniqueness constraint no longer holds SCHEMA lock
while populating the index. From POV of creating the uniqueness constraint
the transaction will still block and potentially fail if the constraint
couldn't be validated on existing data and concurrent updates. The big difference
is that the schema write lock isn't held while populating the constraint index.
This allows other transactions modifying these properties and labels to
progress and complete normally.