parsing of load balancing filter configuration
Filters are configured for the load balancing strategy called
server_policy using a strategy-specific syntax. This syntax allows
specifying a list of rules, each consisting of a chain of filters.

This commit parses that syntax and enables several different policies
with different rulesets to be defined and parsed.

Testing infrastructure has been added to make tests of the functionality
easy and various valid as well as invalid configurations are tested so
that they produce the expected result.

Three filters exist today:
  all: returns all servers
  min: only returns a valid result if the minimum count is met
  tags: only returns servers with matching tags