Fix bug with combined aggregation and sorting
Fixes bug in compiled runtime with combined aggregation and sorting related
to the tracking of projected variables in the context.
Aggregation should only retain any grouping variables, and project
the aggregation variables that it introduces, so that Sort can pick up the
correct set of available variables from the context.