Raise availability guard when core or read replica copy store
Instance might be forced to copy store from another instance if it requires to
pull a transaction log that has already been pruned away. This could happen
when transaction pulling (catch up) experience problems like network outage.
Database is not operational during store copy, most components including the
datasource are shutdown so no transactions can be served.

However it was possible for users (bolt or embedded) to attempt start new
transactions which would hang because of kernel components being stopped.
In similar situations availability guard must be raised to block new
transactions.

This commit makes `LocalDatabase` raise availability guard when it is stopped.
`LocalDatabase` is the component used by cores and read replicas to manage the
lifecycle of the datasource. It is restarted during store copying.